requirements.txt
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
psycopg==3.1.9
pytest==7.4.0
typing_extensions==4.7.1